Code Generation: fix missing include in headers declaring structs
The issue arises when the idl only declares structs, so to test this fix,
it was required to make sure qilang supported to take several idl files
at the same time.

Test IDL files were moved around, and a specific test manages to reproduce
the issue in a separate compilation unit.

Change-Id: I2277e4cbdd98a7fe87a28975a3f56f6b167975e3
Reviewed-on: http://gerrit.aldebaran.lan/69802
Reviewed-by: hcuche <hcuche@aldebaran.com>
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>