Fix importing packages from other projects, including subpackages
In the CMake function, renamed some variables for a better understanding
while adding a new feature: IDL files are installed in the share folder
so that dependent projects can import them.

On the generator side, this IDL files are looked up when referencing
imported packages.

Includes also some refactoring of the qi_gen_idl method, as well as a fix in
generation of header guards, where "." was directly taken from the subpackage
name!

Change-Id: I4ea51a448eb3523057787393fd2ea6b602f77768
Reviewed-on: http://gerrit.aldebaran.lan/69930
Reviewed-by: jlanca <jlanca@aldebaran.com>
Reviewed-by: hcuche <hcuche@aldebaran.com>
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>