IDL parsing: fix error reporting for proper integration in IDEs
The error message was fixed by adding the space in "<path>: error:" and
outputting the data to std::cerr. An extra (annoying) diagnostic message
replication was removed. To allow jumping directly to the error in IDEs,
a symbolic link is used to regroup IDL files in the build directory, on
"UNIX" platforms only (Mac and Linux).

Change-Id: I4c8de4fd152b4874c6c298924c6ae145f952eded
Reviewed-on: http://gerrit.aldebaran.lan/81395
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vpaleologue <vpaleologue@aldebaran-robotics.com>
Reviewed-on: http://gerrit.aldebaran.lan/81813
Reviewed-by: vpaleologue <vpaleologue@aldebaran-robotics.com>