lower qicc's dependency on qi/path.conf
The --target-sdk-dir is used to provide the SDK
directory for the target platform.

qicc uses it to find the qi/path.conf
(typically generated by qibuild) whose content
defines the idl.qi lookup paths.

With this commit, the target-sdk-dir itself is added
to the lookup path (in addition to the qi/path.conf ones).

This makes qicc usable without qi/path.conf, and thus
without qibuild.

Note: maybe we'll need to go further and accept several
      --target-sdk-dir at some point.

Change-Id: I62df13def7ce54cef96f1c92944ce3740aefff58
Reviewed-on: http://gerrit.aldebaran.lan/89159
Reviewed-by: luis.menina <luis.menina@external.softbankrobotics.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: sbarthelemy <sbarthelemy@aldebaran.com>
Reviewed-by: vpaleologue <vpaleologue@aldebaran-robotics.com>
Tested-by: sbarthelemy <sbarthelemy@aldebaran.com>