Fixes generating undefined-behavior code.  (#44282)
Fixes the code generated for member functions of async
types which didn't properly handle the lifetime of the
implementation object to call, resulting in code accessing
an implementation object that was already destroyed.

The issue can be detected when running the test
"QiLangFunction.callsAreSafeWhenDestroyed" added in this patch.
Otherwise it was silent and source of random crashes.

Change-Id: I9b66fcac432dd11f3c0d341a67ef9ca60348f075
Reviewed-on: http://gerrit.aldebaran.lan/97061
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: sbarthelemy <sbarthelemy@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>