Fixes generated code not calling qi::Actor functions safely #45639
Fixes regression introduced by the change with id
    I9b66fcac432dd11f3c0d341a67ef9ca60348f075
that removed the thread-safety guarantee given by the generated code
calling member functions of implementations when the implementation
inherits from `qi::Actor`.
In this specific case the call must be done through the internal strand
of the `qi::Actor`.
This was removed by mistake in a patch that fixed the guarantee that the
member call is done only if the object is still alive.

Therefore, this patch merges the two requirements (see also #44282).
It also adds tests checking our concurrency guarantees for member calls.

The code common to all generated call implementations is injected in
the generated code instead of providing the header and using `#include`.
This is because currently users' targets depending on code generated by
`qicc` do not depend on a common library target, except `libqi`.
We want to avoid having to add in `libqi` code that is specific to
`libqilang`. Instead, `qicc` will inject the code that would have been
included directly at generation-time.
An alternative would have been to make targets depends on a new library
but that would break compatibility with known users setup.

Change-Id: Ibf7a085e4331e5d51ab0e6d9daf40edf86f0b67c
Signed-off-by: A. Joël Lamotte <jlamotte@softbankrobotics.com>
Reviewed-on: http://gerrit.aldebaran.lan/97413
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>