tests: Fixes a possible false negative.
In test `QiLangFunction.callsAreSafeWhenDestroyed`:
`futCall` could end up in error if the object was destroyed before we
entered the `accessMember` function.
In this case we want to get an error, but the test was assuming we would
always get a future not in error. To provide that guarantee in this
patch we make sure that we do not destroy the object before first
entering the beginning of the function and start waiting.
Testing that we actually get an error when the object is destroyed
before entering the function doesn't seem possible without hacks so we
just test for the case where we should never have errors.

Change-Id: Id645c3d185c58c1ee32676cd7f1c46af6a5d10ea
Reviewed-on: http://gerrit.aldebaran.lan/97450
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>