build: Rewrites project build definition. #SW-159
This patch rewrites the definition of the build of qilang in order to
remove the dependency to `qibuild` and `qi toolchains` and instead use
more standard tools and conventions.

It completely refactors the `CMakeLists.txt` files to use an approach
closer to what is referred to as "modern CMake". The project now requires
CMake 3.23 or greater.

The project declares a single target:
- `qilang::qicc`: an executable.

The target is defined for:
- qilang as a CMake subproject (through `add_subdirectory`).
- a build directory of the project.
- an installation of the project.

The last two cases are implemented by generating package configuration
files (`X-config.cmake`) in the build directory and in the installation
directory.

Two package configuration files are created:
- `qilang-config.cmake`.

An additional `qilang-tools-config.cmake` file is generated but it does not
define targets. It instead defines functions for user projects to
generate source files from IDL.

The project defines 2 install components:
- `runtime`, which only installs the `qicc` executable.
- `devel`, which installs files needed for development, which are
  the CMake package configuration files.

When installed, components use the following layout:
- `bin/`: contains the `qicc` executable.
- `lib/cmake`: contains exported CMake package configuration files.

Tests are now declared to CMake using `add_test` and the CTest module is
used. This means that tests can now be executed with the `ctest` command
line tool.

This patch also adds the support of Conan, to download the project
dependencies and to create a Conan package of the project. Please refer
to the README for basic usage.