cmake: Restrict variable substitution in some configured files.
This change prevents some variables substitution in some files configured through the `configure_file` CMake command. In particular, `qilang-tools-config.cmake` was incorrectly configured and unusable once installed.

The `configure_file` command is often used to only copy files in the build directory, and to add a build system dependency to the input file so that it is copied again when changed. In some cases, some variables substitution is required, but we only use the `@VAR@` syntax in the input file for that (instead of the `${VAR}` syntax, which can be ambiguous). Therefore, the calls to `configure_file` are corrected by this change so that either `COPYONLY` or `@ONLY` are used, respectively if we only need a copy or if we need variable substitution.